relocated firestore-types to common directory
